Trait reactive_graph_type_system_api::entity_type_manager::EntityTypeManager

pub trait EntityTypeManager:
    Send
    + Sync
    + Lifecycle {

    // Required methods
    fn register(
        &self,
        entity_type: EntityType,
    ) -> Result<EntityType, EntityTypeRegistrationError>;
    fn get_all(&self) -> EntityTypes;
    fn get_type_ids(&self) -> EntityTypeIds;
    fn get_namespaces(&self) -> Namespaces;
    fn get_by_namespace(&self, namespace: &str) -> EntityTypes;
    fn get_types_by_namespace(&self, namespace: &str) -> EntityTypeIds;
    fn get_by_having_component(
        &self,
        component_ty: &ComponentTypeId,
    ) -> EntityTypes;
    fn has(&self, ty: &EntityTypeId) -> bool;
    fn has_by_type(&self, namespace: &str, type_name: &str) -> bool;
    fn get(&self, ty: &EntityTypeId) -> Option<EntityType>;
    fn get_by_type(
        &self,
        namespace: &str,
        type_name: &str,
    ) -> Option<EntityType>;
    fn find_by_type_name(&self, search: &str) -> EntityTypes;
    fn count(&self) -> usize;
    fn count_by_namespace(&self, namespace: &str) -> usize;
    fn create_entity_type(
        &self,
        ty: &EntityTypeId,
        description: &str,
        components: ComponentTypeIds,
        properties: PropertyTypes,
        extensions: Extensions,
    ) -> Result<EntityType, EntityTypeCreationError>;
    fn update_description(
        &self,
        ty: &EntityTypeId,
        description: &str,
    ) -> Result<EntityType, EntityTypeUpdateError>;
    fn merge(
        &self,
        entity_type_to_merge: EntityType,
    ) -> Result<EntityType, EntityTypeMergeError>;
    fn add_component(
        &self,
        ty: &EntityTypeId,
        component: &ComponentTypeId,
    ) -> Result<(), EntityTypeAddComponentError>;
    fn remove_component(
        &self,
        ty: &EntityTypeId,
        component: &ComponentTypeId,
    ) -> Result<ComponentTypeId, EntityTypeRemoveComponentError>;
    fn add_property(
        &self,
        ty: &EntityTypeId,
        property: PropertyType,
    ) -> Result<PropertyType, EntityTypeAddPropertyError>;
    fn update_property(
        &self,
        ty: &EntityTypeId,
        property_name: &str,
        property_type: PropertyType,
    ) -> Result<PropertyType, EntityTypeUpdatePropertyError>;
    fn remove_property(
        &self,
        ty: &EntityTypeId,
        property_name: &str,
    ) -> Result<PropertyType, EntityTypeRemovePropertyError>;
    fn add_extension(
        &self,
        ty: &EntityTypeId,
        extension: Extension,
    ) -> Result<ExtensionTypeId, EntityTypeAddExtensionError>;
    fn update_extension(
        &self,
        entity_ty: &EntityTypeId,
        extension_ty: &ExtensionTypeId,
        extension: Extension,
    ) -> Result<Extension, EntityTypeUpdateExtensionError>;
    fn remove_extension(
        &self,
        entity_ty: &EntityTypeId,
        extension_ty: &ExtensionTypeId,
    ) -> Result<Extension, EntityTypeRemoveExtensionError>;
    fn delete(&self, ty: &EntityTypeId) -> Option<EntityType>;
    fn validate(&self, ty: &EntityTypeId) -> bool;

}

Required Methods

fn register( &self, entity_type: EntityType, ) -> Result<EntityType, EntityTypeRegistrationError>

fn get_all(&self) -> EntityTypes

Returns all entity types.

fn get_type_ids(&self) -> EntityTypeIds

Returns the ids of all entity types.

fn get_namespaces(&self) -> Namespaces

Returns all defined namespaces.

fn get_by_namespace(&self, namespace: &str) -> EntityTypes

Returns all entity types of the given namespace

fn get_types_by_namespace(&self, namespace: &str) -> EntityTypeIds

Returns all entity types of the given namespace

fn get_by_having_component(&self, component_ty: &ComponentTypeId) -> EntityTypes

Returns all entity types of the given namespace

fn has(&self, ty: &EntityTypeId) -> bool

Returns true, if a entity type with the given name exists.

fn has_by_type(&self, namespace: &str, type_name: &str) -> bool

Returns true, if a entity type with the given fully qualified name exists.

fn get(&self, ty: &EntityTypeId) -> Option<EntityType>

Returns the entity type with the given name or empty.

fn get_by_type(&self, namespace: &str, type_name: &str) -> Option<EntityType>

Returns the entity type with the given fully qualified name or empty.

fn find_by_type_name(&self, search: &str) -> EntityTypes

Returns all entity types whose names matches the given search string.

fn count(&self) -> usize

Returns the count of entity types.

fn count_by_namespace(&self, namespace: &str) -> usize

Returns the count of entity types of the given namespace.

fn create_entity_type( &self, ty: &EntityTypeId, description: &str, components: ComponentTypeIds, properties: PropertyTypes, extensions: Extensions, ) -> Result<EntityType, EntityTypeCreationError>

Creates a new entity type.

fn update_description( &self, ty: &EntityTypeId, description: &str, ) -> Result<EntityType, EntityTypeUpdateError>

Updates the description of the given entity type.

fn merge( &self, entity_type_to_merge: EntityType, ) -> Result<EntityType, EntityTypeMergeError>

Merges the given entity_type_to_merge into an existing entity type with the same entity type id.

fn add_component( &self, ty: &EntityTypeId, component: &ComponentTypeId, ) -> Result<(), EntityTypeAddComponentError>

Adds the component with the given component_name to the entity type with the given name.

fn remove_component( &self, ty: &EntityTypeId, component: &ComponentTypeId, ) -> Result<ComponentTypeId, EntityTypeRemoveComponentError>

Remove the component with the given component_name from the entity type with the given name.

fn add_property( &self, ty: &EntityTypeId, property: PropertyType, ) -> Result<PropertyType, EntityTypeAddPropertyError>

Adds a property to the entity type with the given name.

fn update_property( &self, ty: &EntityTypeId, property_name: &str, property_type: PropertyType, ) -> Result<PropertyType, EntityTypeUpdatePropertyError>

Updates the property with the given property_name. It’s possible to rename the property by using another name in the new property than the provided property_name.

fn remove_property( &self, ty: &EntityTypeId, property_name: &str, ) -> Result<PropertyType, EntityTypeRemovePropertyError>

Removes the property with the given property_name from the entity type with the given name.

fn add_extension( &self, ty: &EntityTypeId, extension: Extension, ) -> Result<ExtensionTypeId, EntityTypeAddExtensionError>

Adds an extension to the entity type with the given name.

fn update_extension( &self, entity_ty: &EntityTypeId, extension_ty: &ExtensionTypeId, extension: Extension, ) -> Result<Extension, EntityTypeUpdateExtensionError>

Updates the property with the given property_name. It’s possible to rename the property by using another name in the new property than the provided property_name.

fn remove_extension( &self, entity_ty: &EntityTypeId, extension_ty: &ExtensionTypeId, ) -> Result<Extension, EntityTypeRemoveExtensionError>

Removes the extension with the given extension_name from the entity type with the given name.

fn delete(&self, ty: &EntityTypeId) -> Option<EntityType>

Deletes the entity type with the given name.

fn validate(&self, ty: &EntityTypeId) -> bool

Validates the entity type with the given name. Tests that all components exists.