Trait reactive_graph_type_system_api::flow_type_manager::FlowTypeManager

pub trait FlowTypeManager:
    Send
    + Sync
    + Lifecycle {

    // Required methods
    fn register(
        &self,
        flow_type: FlowType,
    ) -> Result<FlowType, FlowTypeRegistrationError>;
    fn get_all(&self) -> FlowTypes;
    fn get_type_ids(&self) -> FlowTypeIds;
    fn get_namespaces(&self) -> Namespaces;
    fn get_by_namespace(&self, namespace: &str) -> FlowTypes;
    fn get_types_by_namespace(&self, namespace: &str) -> FlowTypeIds;
    fn has(&self, ty: &FlowTypeId) -> bool;
    fn has_by_type(&self, namespace: &str, type_name: &str) -> bool;
    fn get(&self, ty: &FlowTypeId) -> Option<FlowType>;
    fn get_by_type(&self, namespace: &str, type_name: &str) -> Option<FlowType>;
    fn find_by_type_name(&self, search: &str) -> FlowTypes;
    fn count(&self) -> usize;
    fn count_by_namespace(&self, namespace: &str) -> usize;
    fn create_flow_type(
        &self,
        ty: &FlowTypeId,
        description: &str,
        wrapper_entity_instance: EntityInstance,
        entity_instances: EntityInstances,
        relation_instances: RelationInstances,
        variables: PropertyTypes,
        extensions: Extensions,
    ) -> Result<FlowType, FlowTypeCreationError>;
    fn update_description(
        &self,
        ty: &FlowTypeId,
        description: &str,
    ) -> Result<FlowType, FlowTypeUpdateError>;
    fn add_entity_instance(
        &self,
        ty: &FlowTypeId,
        entity_instance: EntityInstance,
    ) -> Result<(), FlowTypeAddEntityInstanceError>;
    fn update_entity_instance(
        &self,
        ty: &FlowTypeId,
        id: Uuid,
        entity_instance: EntityInstance,
    ) -> Result<(Uuid, EntityInstance), FlowTypeUpdateEntityInstanceError>;
    fn remove_entity_instance(
        &self,
        ty: &FlowTypeId,
        id: Uuid,
    ) -> Result<Option<(Uuid, EntityInstance)>, FlowTypeRemoveEntityInstanceError>;
    fn add_relation_instance(
        &self,
        ty: &FlowTypeId,
        relation_instance: RelationInstance,
    ) -> Result<(), FlowTypeAddRelationInstanceError>;
    fn update_relation_instance(
        &self,
        ty: &FlowTypeId,
        id: &RelationInstanceId,
        relation_instance: RelationInstance,
    ) -> Result<(RelationInstanceId, RelationInstance), FlowTypeUpdateRelationInstanceError>;
    fn remove_relation_instance(
        &self,
        ty: &FlowTypeId,
        id: &RelationInstanceId,
    ) -> Result<Option<(RelationInstanceId, RelationInstance)>, FlowTypeRemoveRelationInstanceError>;
    fn add_extension(
        &self,
        ty: &FlowTypeId,
        extension: Extension,
    ) -> Result<ExtensionTypeId, FlowTypeAddExtensionError>;
    fn update_extension(
        &self,
        flow_ty: &FlowTypeId,
        extension_ty: &ExtensionTypeId,
        extension: Extension,
    ) -> Result<Extension, FlowTypeUpdateExtensionError>;
    fn remove_extension(
        &self,
        flow_ty: &FlowTypeId,
        extension_ty: &ExtensionTypeId,
    ) -> Result<Extension, FlowTypeRemoveExtensionError>;
    fn add_variable(
        &self,
        ty: &FlowTypeId,
        variable: PropertyType,
    ) -> Result<Variable, FlowTypeAddVariableError>;
    fn update_variable(
        &self,
        ty: &FlowTypeId,
        variable_name: &str,
        variable: PropertyType,
    ) -> Result<Variable, FlowTypeUpdateVariableError>;
    fn remove_variable(
        &self,
        ty: &FlowTypeId,
        variable_name: &str,
    ) -> Result<Variable, FlowTypeRemoveVariableError>;
    fn delete(&self, ty: &FlowTypeId) -> Option<FlowType>;
    fn validate(&self, ty: &FlowTypeId) -> bool;

}

Required Methods

fn register( &self, flow_type: FlowType, ) -> Result<FlowType, FlowTypeRegistrationError>

fn get_all(&self) -> FlowTypes

Returns all flow types.

fn get_type_ids(&self) -> FlowTypeIds

Returns the ids of all flow types.

fn get_namespaces(&self) -> Namespaces

Returns all defined namespaces.

fn get_by_namespace(&self, namespace: &str) -> FlowTypes

Returns all flow types.

fn get_types_by_namespace(&self, namespace: &str) -> FlowTypeIds

fn has(&self, ty: &FlowTypeId) -> bool

Returns true, if a flow type with the given name exists.

fn has_by_type(&self, namespace: &str, type_name: &str) -> bool

Returns true, if a flow type with the given fully qualified name exists.

fn get(&self, ty: &FlowTypeId) -> Option<FlowType>

Returns the flow type with the given name or empty.

fn get_by_type(&self, namespace: &str, type_name: &str) -> Option<FlowType>

Returns the flow type with the given fully qualified name or empty.

fn find_by_type_name(&self, search: &str) -> FlowTypes

Returns all flow types whose names matches the given search string.

fn count(&self) -> usize

Returns the count of flow types.

fn count_by_namespace(&self, namespace: &str) -> usize

Returns the count of flow types of the given namespace.

fn create_flow_type( &self, ty: &FlowTypeId, description: &str, wrapper_entity_instance: EntityInstance, entity_instances: EntityInstances, relation_instances: RelationInstances, variables: PropertyTypes, extensions: Extensions, ) -> Result<FlowType, FlowTypeCreationError>

Creates a new flow type.

fn update_description( &self, ty: &FlowTypeId, description: &str, ) -> Result<FlowType, FlowTypeUpdateError>

Updates the description of the given flow type.

fn add_entity_instance( &self, ty: &FlowTypeId, entity_instance: EntityInstance, ) -> Result<(), FlowTypeAddEntityInstanceError>

Adds the given entity instance to the given flow type.

fn update_entity_instance( &self, ty: &FlowTypeId, id: Uuid, entity_instance: EntityInstance, ) -> Result<(Uuid, EntityInstance), FlowTypeUpdateEntityInstanceError>

Updates the entity instance with the given id of the given flow type.

fn remove_entity_instance( &self, ty: &FlowTypeId, id: Uuid, ) -> Result<Option<(Uuid, EntityInstance)>, FlowTypeRemoveEntityInstanceError>

Removes the entity instance with the given id from the given flow type.

fn add_relation_instance( &self, ty: &FlowTypeId, relation_instance: RelationInstance, ) -> Result<(), FlowTypeAddRelationInstanceError>

Adds the given relation instance to the given flow type.

fn update_relation_instance( &self, ty: &FlowTypeId, id: &RelationInstanceId, relation_instance: RelationInstance, ) -> Result<(RelationInstanceId, RelationInstance), FlowTypeUpdateRelationInstanceError>

Updates the relation instance with the given id of the given flow type.

fn remove_relation_instance( &self, ty: &FlowTypeId, id: &RelationInstanceId, ) -> Result<Option<(RelationInstanceId, RelationInstance)>, FlowTypeRemoveRelationInstanceError>

Removes the relation instance with the given id from the given flow type.

fn add_extension( &self, ty: &FlowTypeId, extension: Extension, ) -> Result<ExtensionTypeId, FlowTypeAddExtensionError>

Adds the given extension to the given flow type.

fn update_extension( &self, flow_ty: &FlowTypeId, extension_ty: &ExtensionTypeId, extension: Extension, ) -> Result<Extension, FlowTypeUpdateExtensionError>

Updates the extension with the given name of the given flow type.

fn remove_extension( &self, flow_ty: &FlowTypeId, extension_ty: &ExtensionTypeId, ) -> Result<Extension, FlowTypeRemoveExtensionError>

Removes the extension with the given name from the given flow type.

fn add_variable( &self, ty: &FlowTypeId, variable: PropertyType, ) -> Result<Variable, FlowTypeAddVariableError>

Adds the given variable to the given flow type.

fn update_variable( &self, ty: &FlowTypeId, variable_name: &str, variable: PropertyType, ) -> Result<Variable, FlowTypeUpdateVariableError>

Updates the variable with the given name of the given flow type.

fn remove_variable( &self, ty: &FlowTypeId, variable_name: &str, ) -> Result<Variable, FlowTypeRemoveVariableError>

Removes the variable with the given name from the given flow type.

fn delete(&self, ty: &FlowTypeId) -> Option<FlowType>

Deletes the given flow type.

fn validate(&self, ty: &FlowTypeId) -> bool

Validates the given flow type. Tests that all entity types and relation types exists and are valid.