Trait reactive_graph_type_system_api::relation_type_manager::RelationTypeManager

pub trait RelationTypeManager: Send + Sync + Lifecycle {

    // Required methods
    fn register(
        &self,
        relation_type: RelationType,
    ) -> Result<RelationType, RelationTypeRegistrationError>;
    fn get_all(&self) -> RelationTypes;
    fn get_type_ids(&self) -> RelationTypeIds;
    fn get_namespaces(&self) -> Namespaces;
    fn get_by_namespace(&self, namespace: &str) -> RelationTypes;
    fn get_types_by_namespace(&self, namespace: &str) -> RelationTypeIds;
    fn get_by_having_component(
        &self,
        component_ty: &ComponentTypeId,
    ) -> RelationTypes;
    fn get_outbound_relation_types(
        &self,
        outbound_ty: &ComponentOrEntityTypeId,
        wildcard: bool,
    ) -> RelationTypes;
    fn get_inbound_relation_types(
        &self,
        inbound_ty: &ComponentOrEntityTypeId,
        wildcard: bool,
    ) -> RelationTypes;
    fn has(&self, ty: &RelationTypeId) -> bool;
    fn has_by_type(&self, namespace: &str, type_name: &str) -> bool;
    fn get(&self, ty: &RelationTypeId) -> Option<RelationType>;
    fn get_by_type(
        &self,
        namespace: &str,
        type_name: &str,
    ) -> Option<RelationType>;
    fn find_by_type_name(&self, search: &str) -> RelationTypes;
    fn count(&self) -> usize;
    fn count_by_namespace(&self, namespace: &str) -> usize;
    fn create_relation_type(
        &self,
        outbound_type: &ComponentOrEntityTypeId,
        ty: &RelationTypeId,
        inbound_type: &ComponentOrEntityTypeId,
        description: &str,
        components: ComponentTypeIds,
        properties: PropertyTypes,
        extensions: Extensions,
    ) -> Result<RelationType, RelationTypeCreationError>;
    fn merge(
        &self,
        relation_type_to_merge: RelationType,
    ) -> Result<RelationType, RelationTypeMergeError>;
    fn add_component(
        &self,
        ty: &RelationTypeId,
        component_ty: &ComponentTypeId,
    ) -> Result<(), RelationTypeAddComponentError>;
    fn remove_component(
        &self,
        ty: &RelationTypeId,
        component_ty: &ComponentTypeId,
    ) -> Result<ComponentTypeId, RelationTypeRemoveComponentError>;
    fn add_property(
        &self,
        ty: &RelationTypeId,
        property: PropertyType,
    ) -> Result<PropertyType, RelationTypeAddPropertyError>;
    fn update_property(
        &self,
        relation_ty: &RelationTypeId,
        property_name: &str,
        property_type: PropertyType,
    ) -> Result<PropertyType, RelationTypeUpdatePropertyError>;
    fn remove_property(
        &self,
        ty: &RelationTypeId,
        property_name: &str,
    ) -> Result<PropertyType, RelationTypeRemovePropertyError>;
    fn add_extension(
        &self,
        ty: &RelationTypeId,
        extension: Extension,
    ) -> Result<ExtensionTypeId, RelationTypeAddExtensionError>;
    fn update_extension(
        &self,
        relation_ty: &RelationTypeId,
        extension_ty: &ExtensionTypeId,
        extension: Extension,
    ) -> Result<Extension, RelationTypeUpdateExtensionError>;
    fn remove_extension(
        &self,
        relation_ty: &RelationTypeId,
        extension_ty: &ExtensionTypeId,
    ) -> Result<Extension, RelationTypeRemoveExtensionError>;
    fn delete(&self, ty: &RelationTypeId) -> Option<RelationType>;
    fn validate(&self, ty: &RelationTypeId) -> bool;

}

Required Methods

fn register( &self, relation_type: RelationType, ) -> Result<RelationType, RelationTypeRegistrationError>

fn get_all(&self) -> RelationTypes

Returns all relation types.

fn get_type_ids(&self) -> RelationTypeIds

Returns the ids of all relation types.

fn get_namespaces(&self) -> Namespaces

Returns all defined namespaces.

fn get_by_namespace(&self, namespace: &str) -> RelationTypes

Returns all relation types of the given namespace

fn get_types_by_namespace(&self, namespace: &str) -> RelationTypeIds

Returns all relation types of the given namespace

fn get_by_having_component( &self, component_ty: &ComponentTypeId, ) -> RelationTypes

Returns all relation types of the given namespace

fn get_outbound_relation_types( &self, outbound_ty: &ComponentOrEntityTypeId, wildcard: bool, ) -> RelationTypes

Returns outbound relation types for the given entity type.

fn get_inbound_relation_types( &self, inbound_ty: &ComponentOrEntityTypeId, wildcard: bool, ) -> RelationTypes

Returns inbound relation types for the given entity type.

fn has(&self, ty: &RelationTypeId) -> bool

Returns true, if a relation type with the given name exists.

fn has_by_type(&self, namespace: &str, type_name: &str) -> bool

Returns true, if a relation type with the given fully qualified name exists.

fn get(&self, ty: &RelationTypeId) -> Option<RelationType>

Returns the relation type with the given name.

fn get_by_type(&self, namespace: &str, type_name: &str) -> Option<RelationType>

Returns the relation type with the given fully qualified name.

fn find_by_type_name(&self, search: &str) -> RelationTypes

Returns all relation types whose names matches the given search string.

fn count(&self) -> usize

Returns the count of relation types.

fn count_by_namespace(&self, namespace: &str) -> usize

Returns the count of relation types of the given namespace.

fn create_relation_type( &self, outbound_type: &ComponentOrEntityTypeId, ty: &RelationTypeId, inbound_type: &ComponentOrEntityTypeId, description: &str, components: ComponentTypeIds, properties: PropertyTypes, extensions: Extensions, ) -> Result<RelationType, RelationTypeCreationError>

Creates a new relation type.

fn merge( &self, relation_type_to_merge: RelationType, ) -> Result<RelationType, RelationTypeMergeError>

Merges the given relation_type_to_merge into an existing relation type with the same relation type id.

fn add_component( &self, ty: &RelationTypeId, component_ty: &ComponentTypeId, ) -> Result<(), RelationTypeAddComponentError>

Adds the component with the given component_name to the relation type with the given name.

fn remove_component( &self, ty: &RelationTypeId, component_ty: &ComponentTypeId, ) -> Result<ComponentTypeId, RelationTypeRemoveComponentError>

Remove the component with the given component_name from the relation type with the given name.

fn add_property( &self, ty: &RelationTypeId, property: PropertyType, ) -> Result<PropertyType, RelationTypeAddPropertyError>

Adds a property to the relation type with the given name.

fn update_property( &self, relation_ty: &RelationTypeId, property_name: &str, property_type: PropertyType, ) -> Result<PropertyType, RelationTypeUpdatePropertyError>

Updates the property with the given property_name. It’s possible to rename the property by using another name in the new property than the provided property_name.

fn remove_property( &self, ty: &RelationTypeId, property_name: &str, ) -> Result<PropertyType, RelationTypeRemovePropertyError>

Removes the property with the given property_name from the relation type with the given name.

fn add_extension( &self, ty: &RelationTypeId, extension: Extension, ) -> Result<ExtensionTypeId, RelationTypeAddExtensionError>

Adds an extension to the relation type with the given name.

fn update_extension( &self, relation_ty: &RelationTypeId, extension_ty: &ExtensionTypeId, extension: Extension, ) -> Result<Extension, RelationTypeUpdateExtensionError>

Updates the extension with the given extension type. It’s possible to rename the extension by using another extension type in the new extension than the provided extension type.

fn remove_extension( &self, relation_ty: &RelationTypeId, extension_ty: &ExtensionTypeId, ) -> Result<Extension, RelationTypeRemoveExtensionError>

Removes the extension with the given extension_name from the relation type with the given name.

fn delete(&self, ty: &RelationTypeId) -> Option<RelationType>

Deletes the relation type with the given name.

fn validate(&self, ty: &RelationTypeId) -> bool

Validates the relation type with the given name. Tests that all components, the outbound and inbound entity type exists.